MAJ add_sonde.sh

+++ scripts/add_sonde.sh
@@ -1,49 +1,74 @@
 #!/bin/bash
 
 echo -e "Script pour l'ajout de nouvelles sondes perfSONAR à l'Observatoire ...\n"
-x=0
-f=0
+s=false ; f=false ; ver="7"
+declare -A startServices=( ['0']="pscheduler-scheduler" ['1']="pscheduler-runner" ['2']="pscheduler-ticker" ['3']="pscheduler-archiver" ['4']="owamp-server" ['5']="bwctl-server" )
 
-usage() { echo "Usage : $0 -s <address> -f <fichier.JSON>" 1>&2; exit 1;}
+function aide { 
+   echo ""
+   echo -e "Usage : \n$0 -s <address> -f <fichier.JSON>" 1>&2; 
+   echo ""
+   echo "-s : L'addresse du serveur perfSONAR"
+   echo "-f : Le nom du fichier JSON"
+   echo ""
+}   
 
-while getopts "s:f:" option; do
-  case "${option}" in
+if [ "$UID" -ne "0" ] ; then
+   echo -e "Vous devez être superutilisateur pour exécuter $0. \nEssayez avec sudo $0";
+   exit 9
+fi
+
+while getopts "s:f:" opts; do
+  case $opts in
     s)
-      x=1
+      s=true
       SERVER=${OPTARG}
       ;;
     f)
-      f=1
+      f=true
       FICHIER=${OPTARG}
       ;;
-    *)
-      usage
+    \?)
+      aide
+      exit 1
       ;;
   esac
 done
 
-function paquets {
-yum -y update
-yum -y install epel-release
-yum -y install http://software.internet2.edu/rpms/el7/x86_64/main/RPMS/perfSONAR-repo-0.8-1.noarch.rpm
-yum clean all
-yum -y install perfsonar-toolkit
-sleep 2
-systemctl start pscheduler-scheduler
-systemctl start pscheduler-runner
-systemctl start pscheduler-archiver
-systemctl start pscheduler-ticker
-systemctl start perfsonar-lsregistrationdaemon
-systemctl start bwctl-server
-systemctl start owamp-server
+#if [[ $(lsb_release -d | cut -f 2 | cut -d" " -f 3) =~ "\$9" ]]
+
+function demarrerServices {
+
+  for s in "${startServices[@]}"
+    service $i start
+  done
+  
 }
 
-ip="^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$"
+function paquets {
+
+  repActuel=$(pwd)
+
+  if [[ -e "/etc/debian_version" ]]; then
+    cd /etc/apt/sources.list.d/
+    wget http://downloads.perfsonar.net/debian/perfsonar-jessie-release.list
+    wget -qO - http://downloads.perfsonar.net/debian/perfsonar-debian-official.gpg.key | apt-key add -
+    cd $repActuel
+    apt-get -y update
+    apt-get -y install perfsonar-toolkit
+  elif [[ -e "/etc/centos-release" ]]; then
+    yum -y install epel-release
+    yum -y install http://software.internet2.edu/rpms/el7/x86_64/main/RPMS/perfSONAR-repo-0.8-1.noarch.rpm
+    yum clean all
+    yum -y install perfsonar-toolkit
+  fi
+  demarrerServices
+}
+
+#ip="^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$"
 
 
-if [ $x -eq 1 ] && [ $f -eq 1 ] ; then
-
-if [[ $SERVER =~ $ip ]] ; then
+if [ $x ] && [ $s ] ; then
 
 reponse=2
 while [ $reponse -ne 1 ] && [ $reponse -ne 0 ] ; do
@@ -55,47 +80,44 @@ if [ $reponse -eq 0 ] ; then
   paquets
 fi
 
-  if ping -c 1 $SERVER &> /dev/null ; then 
-    if curl -f http://$SERVER/$FICHIER ; then
-      echo -e "\n*******************************************************************\nFichier JSON trouvé ..."
-      sleep 2
-      echo -e "\n\nINFORMATION: UBICATION DE LA SONDE. Entrez la ville :"
-      read city
-      echo "INFORMATION: UBICATION DE LA SONDE. Entrez l'état :"
-      read state
-      echo "INFORMATION: UBICATION DE LA SONDE. Entrez la latitude :"
-      read lati
-      echo "INFORMATION: UBICATION DE LA SONDE. Entrez la longitude :"
-      read longi
-      echo "INFORMATION: Entrez une description :"
-      read descr
-      echo "INFORMATION: Entrez l'addresse de la sonde :"
-      read addr
+if ping -c 1 $SERVER &> /dev/null ; then 
+  if curl -f http://$SERVER/$FICHIER ; then
+    echo -e "\n*******************************************************************\nFichier JSON trouvé ..."
+    sleep 2
+    echo -e "\n\nINFORMATION: EMPLACEMENT DE LA SONDE. Entrez la ville :"
+    read city
+    echo "INFORMATION: EMPLACEMENT DE LA SONDE. Entrez l'état :"
+    read state
+    echo "INFORMATION: EMPLACEMENT DE LA SONDE. Entrez la latitude :"
+    read lati
+    echo "INFORMATION: EMPLACEMENT DE LA SONDE. Entrez la longitude :"
+    read longi
+    echo "INFORMATION: Entrez une description :"
+    read descr
+    echo "INFORMATION: Entrez l'addresse de la sonde :"
+    read addr
 
-      echo "ville: "$city"" >> ./data.yaml
-      echo "etat: "$state"" >> ./data.yaml
-      echo "lat: "$lati"" >> ./data.yaml
-      echo "lon: "$longi"" >> ./data.yaml
-      echo "dec: "$descr"" >> ./data.yaml
-      echo "add: "$addr"" >> ./data.yaml
+    echo "ville: "$city"" >> ./data.yaml
+    echo "etat: "$state"" >> ./data.yaml
+    echo "lat: "$lati"" >> ./data.yaml
+    echo "lon: "$longi"" >> ./data.yaml
+    echo "dec: "$descr"" >> ./data.yaml
+    echo "add: "$addr"" >> ./data.yaml
 
-      echo -e "\nAppel au script de modification du fichier mesh config : $FICHIER ..."
-      sleep 2
-      ./maj_mesh-config.py
-      echo -e "\n***********************************************************************\nNettoyage ..."
-      rm -f ./data.yaml
-      sleep 1
-    else
-      echo -e "\nLe fichier entré n'existe pas dans le serveur !"
-    fi
+    echo -e "\nAppel au script de modification du fichier mesh config : $FICHIER ..."
+    sleep 2
+    ./maj_mesh-config.py
+    echo -e "\n***********************************************************************\nNettoyage ..."
+    rm -f ./data.yaml
+    sleep 1
   else
-    echo "Le serveur n'est pas disponible. Veulliez vérifier l'addresse ou essayez plus tard."
+    echo -e "\nLe fichier entré n'existe pas dans le serveur !"
   fi
 else
-  echo "Addresse IP invalide. Veulliez en entrer une autre"
-  exit 2
+  echo "Le serveur n'est pas disponible. Veulliez vérifier l'addresse ou essayez plus tard."
 fi
+
 else
-  echo -e "Vous devez entrer l'addresse du serveur MESH et le nom du fichier JSON.\nUsage : $0 -s <address> -f <fichier.JSON>"
+   aide
 fi
